Fix dependency cycles in `preferences` package
